+++ docsrc/zs/guides/release_notes.rst
@@ -18,6 +18,14 @@ Bug Fixes
 ^^^^^^^^^
 
 - Fixed ZPA Pagination pagesize parameter to the maximum supported of `500`  (`#118 <https://github.com/zscaler/zscaler-sdk-python/pull/118>`_)
+- Fixed ZIA Isolation Profile method misconfiguration (`#118 <https://github.com/zscaler/zscaler-sdk-python/pull/118>`_)
+
+Enhancements
+^^^^^^^^^^^^
+
+- Added the following new ZIA location management endpoints (`#118 <https://github.com/zscaler/zscaler-sdk-python/pull/118>`_)
+    - `locations/bulkDelete`
+    - `locations/groups/count`
 
 0.6.0 (June, 28 2024)
 ----------------------